refactor: fake 객체는 test에 있도록 수정

+++ src/main/java/blackjack/view/output/ConsoleOutputView.java
@@ -47,7 +47,7 @@ public class ConsoleOutputView implements OutputView {
         System.out.println();
     }
 
-    private void printResultOfDealer(double profitOfDealer) {
+    private void printResultOfDealer(int profitOfDealer) {
         System.out.println("딜러: " + profitOfDealer);
     }
